% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllGenerics.R, R/ModifierML-data.R
\docType{methods}
\name{assembleTrainingData}
\alias{assembleTrainingData}
\alias{assembleTrainingData,ModifierML,GRanges-method}
\alias{assembleTrainingData,ModifierML,GRangesList-method}
\title{Assemble training data from aggregate sequence data}
\usage{
assembleTrainingData(x, coord)

\S4method{assembleTrainingData}{ModifierML,GRanges}(x, coord)

\S4method{assembleTrainingData}{ModifierML,GRangesList}(x, coord)
}
\description{
title
}
\examples{
ModifierMLKeras()
}
