% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierMLKeras-class.R
\name{ModifierMLKeras-class}
\alias{ModifierMLKeras-class}
\alias{ModifierMLKeras}
\title{ModifierMLKeras class}
\description{
The \code{ModifierMLKeras} class extends the virtual class
\code{ModifierMLModel} and unifies the access to \code{Keras} machine
learning models used in the detection of post-transcriptional modifications
in RNA sequencing data. The \code{ModifierMLKeras} class is virtual itself
and must be extended from for each individual machine learning model.
}
\seealso{
\code{\link[=ModifierMLModel]{ModifierMLModel}}
}
