% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierMLModel-class.R,
%   R/ModifierMLKeras-class.R
\docType{class}
\name{ModifierMLModel-class}
\alias{ModifierMLModel-class}
\alias{ModifierMLModel}
\alias{ModifierMLKeras-class}
\title{ModifierMLModel virtual class}
\description{
Teh \code{ModifierMLModel} is a virtual class and is used for representing
different types of machine learning models used in the detection of
post transcriptional modifiations in RNA sequencing data.

The next class inheriting from here should only implement a certain type
of model and also be virtual. The grand child of the \code{ModifierMLModel}
class should than implement a specific model for detecting certain types
of modifications.
}
\section{Slots}{

\describe{
\item{\code{model}}{a machine learning object of any type}

\item{\code{modelFile}}{a file representation of the model}
}}

\seealso{
\code{\link[=ModifierMLKeras-class]{ModifierMLKeras}}
}
