% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModifierMLranger-class.R
\docType{class}
\name{ModifierMLranger-class}
\alias{ModifierMLranger-class}
\alias{ModifierMLranger}
\alias{useModel,ModifierMLranger,ModifierML-method}
\title{ModifierMLranger class}
\usage{
\S4method{useModel}{ModifierMLranger,ModifierML}(x, y)
}
\arguments{
\item{x}{a \code{ModifierMLranger} object}

\item{y}{a \code{ModifierML} object}
}
\value{
a \code{ModifierMLranger} object
}
\description{
The \code{ModifierMLranger} class extends the virtual class
\code{ModifierMLModel} and unifies the access to
\code{\link[ranger:ranger]{ranger}} machine learning models used in the
detection of post-transcriptional modifications in RNA sequencing data. The
\code{ModifierMLranger} class is virtual itself and must be extended from for
each individual machine learning model.
}
\examples{
# example class derived from the virtual ModifierMLranger class
setClass("ModifierMLexample",
         contains = c("ModifierMLranger"))
ModifierMLexample <- function(...){
  new("ModifierMLexample")
}
mlmodel <- ModifierMLexample()
}
\seealso{
\code{\link[=ModifierMLModel]{ModifierMLModel}}
}
