% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.ML.R
\docType{package}
\name{RNAmodR.ML}
\alias{RNAmodR.ML}
\alias{RNAmodR.ML-package}
\title{RNAmodR.ML}
\description{

}
\references{
-
}
\seealso{
The \code{RNAmodR} and \code{RNAmodR.m1A} package.
}
\author{
Felix G M Ernst [aut]
}
