% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.ML.R
\docType{data}
\name{RNAmodR.ML-datasets}
\alias{RNAmodR.ML-datasets}
\alias{dmod}
\alias{mod7}
\alias{me}
\alias{model}
\title{Example data in the RNAmodR.ML package}
\format{
a \code{GRanges} object containg D positions

a \code{GRanges} object containg m7G positions

a \code{ModMLExample} object for examples

a \code{ranger} object for examples
}
\usage{
data(dmod)

data(mod7)

data(me)

data(model)
}
\description{
This contains an example ModifierSet object
}
\keyword{datasets}
