% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RNAmodR.ML.R
\docType{package}
\name{RNAmodR.ML}
\alias{RNAmodR.ML}
\alias{RNAmodR.ML-package}
\title{RNAmodR.ML}
\description{
To extend the functionality of the \code{RNAmodR} package and classical
detection strategies towards detection through machine learning models,
\code{RNAmodR.ML} provides classes and an example workflow.
}
\seealso{
\code{\link[RNAmodR:RNAmodR]{RNAmodR}} package.
}
\author{
Felix G M Ernst [aut]
}
